+++ dart_fss/tests/test_case/testcrp.py
@@ -1,7 +1,6 @@
-import re
 import pytest
 import dart_fss.crp as crp
-from dart_fss._utils import compare_str
+from dart_fss.utils._utils import compare_str
 from dart_fss.fs_search import find_all_columns
