+++ dart_fss/utils/singleton.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+class Singleton(type):
+    """
+    Singleton metaclass
+    """
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
